Move news out of associations
