+++ app/src/main/java/be/ugent/zeus/hydra/news/list/NewsItemViewHolder.java
@@ -1,11 +1,11 @@
-package be.ugent.zeus.hydra.association.news.list;
+package be.ugent.zeus.hydra.news.list;
 
 import android.text.TextUtils;
 import android.view.View;
 import android.widget.TextView;
 
 import be.ugent.zeus.hydra.R;
-import be.ugent.zeus.hydra.association.news.UgentNewsArticle;
+import be.ugent.zeus.hydra.news.UgentNewsArticle;
 import be.ugent.zeus.hydra.common.ArticleViewer;
 import be.ugent.zeus.hydra.common.ui.customtabs.ActivityHelper;
 import be.ugent.zeus.hydra.common.ui.html.Utils;
