+++ app/src/main/java/be/ugent/zeus/hydra/news/list/NewsViewModel.java
@@ -1,9 +1,9 @@
-package be.ugent.zeus.hydra.association.news.list;
+package be.ugent.zeus.hydra.news.list;
 
 import android.app.Application;
 
-import be.ugent.zeus.hydra.association.news.UgentNewsArticle;
-import be.ugent.zeus.hydra.association.news.UgentNewsRequest;
+import be.ugent.zeus.hydra.news.UgentNewsArticle;
+import be.ugent.zeus.hydra.news.UgentNewsRequest;
 import be.ugent.zeus.hydra.common.request.Request;
 import be.ugent.zeus.hydra.common.ui.RequestViewModel;
