+++ app/src/test/java/be/ugent/zeus/hydra/association/UgentNewsArticleTest.java
@@ -1,6 +1,6 @@
 package be.ugent.zeus.hydra.association;
 
-import be.ugent.zeus.hydra.association.news.UgentNewsArticle;
+import be.ugent.zeus.hydra.news.UgentNewsArticle;
 import be.ugent.zeus.hydra.common.ModelTest;
 import be.ugent.zeus.hydra.testing.Utils;
 import org.junit.Test;
