+++ app/src/test/java/be/ugent/zeus/hydra/common/ArticleViewerTest.java
@@ -2,7 +2,7 @@ package be.ugent.zeus.hydra.common;
 
 import android.net.Uri;
 
-import be.ugent.zeus.hydra.association.news.UgentNewsArticle;
+import be.ugent.zeus.hydra.news.UgentNewsArticle;
 import be.ugent.zeus.hydra.common.ui.customtabs.ActivityHelper;
 import be.ugent.zeus.hydra.schamper.Article;
 import be.ugent.zeus.hydra.testing.RobolectricUtils;
