+++ app/src/test/java/be/ugent/zeus/hydra/feed/cards/news/NewsItemViewHolderTest.java
@@ -3,7 +3,7 @@ package be.ugent.zeus.hydra.feed.cards.news;
 import android.view.View;
 
 import be.ugent.zeus.hydra.R;
-import be.ugent.zeus.hydra.association.news.UgentNewsArticle;
+import be.ugent.zeus.hydra.news.UgentNewsArticle;
 import be.ugent.zeus.hydra.feed.cards.implementations.AbstractFeedViewHolderTest;
 import org.junit.Test;
 import org.junit.runner.RunWith;
